divide packages
