+++ src/main/java/collections/Queue.java
@@ -1,4 +1,4 @@
-package jalt;
+package jalt.collections;
 
 import java.util.Iterator;
