+++ src/main/java/sorts/QuickSort.java
@@ -1,4 +1,4 @@
-package jalt;
+package jalt.sorts;
 
 import java.util.Comparator;
 import static jalt.Utils.isLessThan;
