+++ src/main/java/sorts/ShellSort.java
@@ -1,4 +1,4 @@
-package jalt;
+package jalt.sorts;
 
 import java.util.Comparator;
 import static jalt.Utils.isLessThanOrEqualTo;
