+++ src/test/java/sorts/InsertionSortTest.java
@@ -1,14 +1,14 @@
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import org.junit.jupiter.api.Test;
 import java.util.Arrays;
-import jalt.SelectionSort;
-import jalt.Sorter;
+import jalt.sorts.InsertionSort;
+import jalt.sorts.Sorter;
 
-public class SelectionSortTest {
+public class InsertionSortTest {
   @Test
   public void testSort() {
     Integer[] toSort = { 3, 1, 2, 4 };
-    Sorter.sort(toSort, new SelectionSort());
+    Sorter.sort(toSort, new InsertionSort());
 
     Integer[] expected = { 1, 2, 3, 4 };
     assertArrayEquals(toSort, expected);
