+++ src/test/java/sorts/MergeSortTest.java
@@ -1,14 +1,14 @@
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import org.junit.jupiter.api.Test;
 import java.util.Arrays;
-import jalt.ShellSort;
-import jalt.Sorter;
+import jalt.sorts.MergeSort;
+import jalt.sorts.Sorter;
 
-public class ShellSortTest {
+public class MergeSortTest {
   @Test
   public void testSort() {
     Integer[] toSort = { 3, 1, 2, 4 };
-    Sorter.sort(toSort, new ShellSort());
+    Sorter.sort(toSort, new MergeSort());
 
     Integer[] expected = { 1, 2, 3, 4 };
     assertArrayEquals(toSort, expected);
